+++ src/lib/config.ts
@@ -60,8 +60,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.0.1-alpha.16",
-  genVersion: "2.811.0",
+  sdkVersion: "0.0.1-alpha.17",
+  genVersion: "2.822.1",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.1-alpha.16 2.811.0 1.0.0 @gbg/go-core",
+    "speakeasy-sdk/typescript 0.0.1-alpha.17 2.822.1 1.0.0 @gbg/go-core",
 } as const;